+++ console/views.py
@@ -27,7 +27,6 @@ PATHS = {
     "receipt": settings.SDX_FTP_RECEIPT_PATH
 }
 
-logging.basicConfig(level=settings.LOGGING_LEVEL, format=settings.LOGGING_FORMAT)
 logger = wrap_logger(logging.getLogger(__name__))