+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   - Add all environment variables to README
   - Remove secure data from logging messages
   - Add license
+  - Adding sdx-common functionality
+  - Updating logger format using sdx-common
 
 ### 1.0.0 2016-08-16
   - Initial release